% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeShear.R
\name{compositeShear}
\alias{compositeShear}
\title{Shear modulus of 2-phase composites using DEM.}
\usage{
compositeShear(k1, mu1, k2, mu2, asp, phic, phi)
}
\arguments{
\item{k1}{Bulk modulus of phase 1.}

\item{mu1}{Shear modulus of phase 1.}

\item{k2}{Bulk modulus of phase 2.}

\item{mu2}{Shear modulus of phase 2.}

\item{asp}{Aspect ratio of inclusions phase 2 in a matrix of phase 1. 'asp'< 1 implies oblate inclusions, 
'asp' = 1 implies spherical inclusions and 'asp' > 1 implies prolate inclusions.}

\item{phic}{Critical porosity/volume fraction of phase 2, endpoint for phase 2 scaled from this value.}

\item{phi}{Porosity/volume fraction of phase 2.}
}
\value{
a named dataframe whose first column is the composite shear modulus computed by DEM, each of the following 
columns are one of the 7 inputs that went into the model after passing the model check.
Note: Instead of 'k1' and 'k2', we return 'k1' and 'conK' = k1/k2, the contrast in bulk modulus between the phases. 
Similarly, instead of 'mu1' and 'mu2', we return 'mu1' and 'conMu' = mu1/mu2.
}
\description{
Generates the effective shear modulus of a mix of spheroidal inclusions of phase 2 (with a specific aspect ratio)
in a matrix of phase 1 using the differential effective medium (DEM) theory.
}
\examples{
k1<- (rnorm(100, 36, 10))
mu1<-(rnorm(100, 45, 10))
k2<- (rnorm(100, 36, 10))
mu2<- (rnorm(100, 45, 10))
asp<- (rnorm(100, 1, .5))
phic<- (rnorm(100, .4, .1))
phi<- (rnorm(100, .25, .1))
df <- compositeShear(k1,mu1,k2,mu2,asp,phic,phi)

}
