% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Data for shearRange or compositeShear modelling}
\usage{
generateData(datalength = 1000, modelfn = shearRange, userdata = NULL,
  v_mean = list(36, 36, 1, 1, 0.3), v_sd = list(10, 10, 0.4, 0.4, 0.2),
  visualize = FALSE)
}
\arguments{
\item{datalength}{The input datalength. Default = 1000}

\item{modelfn}{The model function we want to use. Default is \code{shearRange}. 
Another function in the library is \code{compositeShear}. The number of arguments
can be found by doing help(\code{compositeShear}). User defined function can also 
be passed. Make sure the to pass the function name without quotes.}

\item{userdata}{User can pass a dataframe with the input variables to the models.
Default is NULL.}

\item{v_mean}{A list of mean values for the inputs. Default is list of
# input means for \code{shearRange}.}

\item{v_sd}{A list of corresponding sd values for the inputs. Default is a 
a list of input sds for \code{shearRange}}

\item{visualize}{If true will create a scatterplot of the input data after
applying all the conditions specified in the geophysical models. Default is false.}
}
\value{
a dataframe with output results and input 
variables. It also uses pair to make a scatterplot of the inputs
}
\description{
Function to generate data give datalength and modelfn 
= \code{shearRange}, \code{compositeShear}. It calculates the
output using geophysical modeling and returns a dataframe.
}
\examples{
df <- generateData()



}
