% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{modelFitting}
\alias{modelFitting}
\title{Variable importance using statistical learning}
\usage{
modelFitting(df, modelFit = "lm", plotPredImp = FALSE)
}
\arguments{
\item{df}{dataframe whose 1st column is the response to be analysed followed by all predictors to be tested}

\item{modelFit}{the statistical model used for fitting.
Values can be "lm", "gb", "rf" and "all". If any other value is given
it will give an error.
Default is lm.}

\item{plotPredImp}{if TRUE it will create a predictor importance plot for the input variables
Default is FALSE.}
}
\value{
a list of test data, fit, predicted values and 
test RMSE of the modeling function.
}
\description{
Given a dataframe produced of response and predictors, this function first 
divides the data into training and test sets. It does model fitting and returns the key results, and optionally, plots,
for further analysis by the user. The statistical methods used are linear regression (lm), random forest (rf) and
gradient boosting (gb). A fourth argument "all" can be passed to return all 3 results.
}
\examples{
c33 <- rnorm(3000, 36, 10)
c44 <- rnorm(3000, 36, 10)
epsilon <- rnorm(3000, 1, 0.4)
gamma   <- rnorm(3000, 1, 0.4)
delta <- rnorm(3000, 0.3, 0.2)
df <- shearRange(c33, c44, epsilon, gamma, delta)
lst <- modelFitting(df, modelFit = "all")


}
