% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shearRange.R
\name{shearRange}
\alias{shearRange}
\title{Voigt-Reuss range (normailsed) for shear modulus}
\usage{
shearRange(c33, c44, epsilon, gamma, delta)
}
\arguments{
\item{c33}{P-wave modulus along axis of symmetry (VTI domain).}

\item{c44}{S-wave modulus along axis of symmetry (VTI domain).}

\item{epsilon}{Thomsen's anisotropy parameter (VTI domain).}

\item{gamma}{Thomsen's anisotropy parameter (VTI domain).}

\item{delta}{Thomsen's anisotropy parameter (VTI domain).}
}
\value{
a dataframe whose first column is the model output (dfg) and the other columns are the input variables that went into the model.
}
\description{
Generates macro-scale isotropic Voigt-Reuss range (normailsed) for shear modulus (dfg) of a micro-scale anisotropic phase with VTI 
(vertical transverse isotropy)symmetry, Many instances of the micro-scale anisotropic domain oriented without directional preference give rise to 
the macro-scale isotropic composite.
}
\examples{
c33 <- rnorm(1000, 36, 10)
c44 <- rnorm(1000, 36, 10)
epsilon <- rnorm(1000, 1, 0.4)
gamma   <- rnorm(1000, 1, 0.4)
delta <- rnorm(1000, 0.3, 0.2)
df <- shearRange(c33, c44, epsilon, gamma, delta)

}
