% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCompositeShear.R
\name{shinyCompositeShear}
\alias{shinyCompositeShear}
\title{Run the shiny app for compositeShear modelling}
\usage{
shinyCompositeShear()
}
\description{
Runs a shiny app compositeShear modelling with
3 statistical methods - lm, random forest and gradient boosting.
}
