% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyShearRange.R
\name{shinyShearRange}
\alias{shinyShearRange}
\title{Run the shiny app for shearRange modelling}
\usage{
shinyShearRange()
}
\description{
Runs a shiny app shearRange modelling with
3 statistical methods - lm, random forest and gradient boosting.
}
