% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgeophy.R
\docType{package}
\name{statgeophy}
\alias{statgeophy}
\alias{statgeophy-package}
\title{Geophysical modeling and statistical fitting}
\description{
It has 2 functions, shearRange and compositeShear. 
Each function uses a different
algorithm to produce a dataframe with calculated output
as the first column and the all the inputs as the other 
columns.
statgeophy::shinyShearRange() to see the shearRange
results.
statgeophy::shinyCompositeShear() shows the compositeShear
results. This one takes a little time to do the calculations.
}
\details{
The \code{generateData} is provided to generate rnorm
data and feed it to the above modelling functions to get the output.

The \code{modelFitting} uses the output from \code{generateData}
to use linear regression, random forest and or boosting to fit a model
and optionally plot the predictor importance.

Two shiny apps \code{shinyShearRange} and \code{shinyCompositeShear}
are provided to show the results of model fitting on a webpage.
}
\references{
Rock Physics Handbook
}
\seealso{
\code{\link{generateData}}

\code{\link{modelFitting}}

\code{\link{shearRange}}

\code{\link{compositeShear}}
}
